requirements.txt
litellm==1.55.3
pydantic==2.10.1
PyYAML==6.0.2
python-dotenv==1.0.1
jsonschema==4.23.0
rich==13.9.4
tiktoken==0.8.0
pytz==2024.2
